refactor: remove useless broker methods
